pyproject.toml
[project]
name = "starbucks-menu-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.75.0",
    "fastapi>=0.128.0",
    "httpx[socks]>=0.28.1",
    "jinja2>=3.1.6",
    "numpy>=2.4.0",
    "openai>=2.14.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.21",
    "scikit-learn>=1.8.0",
    "uvicorn>=0.40.0",
]
